Factor out user space symbolization logic into Usyms type
Move all user space address symbolization logic into the newly created
Usyms type. Having a type dedicated to the symbolization process makes
for cleaner separation of concerns and will allow us to more easily swap
out the implementation.
